+++ backend/app/project/api/helpers/location_checks.py
@@ -113,7 +113,7 @@ class AbstractLocationActionValidator(abc.ABC):
         The begin date must be there for the creation while the
         end date is optional.
         """
-        if "begin_date" not in payload_dict["attributes"].keys():
+        if "begin_date" not in payload_dict.get("attributes", {}).keys():
             raise BadRequestError("begin_date is required.")
         begin_date_as_string = payload_dict["attributes"]["begin_date"]
         end_date_as_string = payload_dict["attributes"].get("end_date")
@@ -180,12 +180,12 @@ class AbstractLocationActionValidator(abc.ABC):
 
     def _extract_updated_begin_and_end_dates(self, payload_dict, existing_location):
         """Extract the begin and end dates for the update case."""
-        if "begin_date" in payload_dict["attributes"].keys():
+        if "begin_date" in payload_dict.get("attributes", {}).keys():
             begin_date_as_string = payload_dict["attributes"]["begin_date"]
             begin_date = self._parse_datetime(begin_date_as_string)
         else:
             begin_date = existing_location.begin_date
-        if "end_date" in payload_dict["attributes"].keys():
+        if "end_date" in payload_dict.get("attributes", {}).keys():
             end_date_as_string = payload_dict["attributes"]["end_date"]
             if end_date_as_string is not None:
                 end_date = self._parse_datetime(end_date_as_string)